+++ backend/models/userModel.js
@@ -1,0 +1,11 @@
+const db = require("../firebase");
+
+exports.getUser = async (email) => {
+    const snapshot = await db.collection("user").where("email", "==", email).get();
+    if (snapshot.empty) return null;
+    return snapshot.docs[0].data();
+}
+
+exports.createUser = async (user) => {
+  return await db.collection("user").add(user);
+};
