+++ backend/utils/token.js
@@ -1,0 +1,5 @@
+const jwt = require("jsonwebtoken");
+
+exports.generateToken = (payload) => {
+  return jwt.sign(payload, process.env.JWT_SECRET, { expiresIn: "7d" });
+};
